Move dataframe structure asserters to utils file
